Python/help/eyetracker/requirements.txt
django>=4.2.0
opencv-python>=4.7.0
numpy>=1.24.0
scipy>=1.10.0
pillow>=9.5.0 